R Programming - Week2 - Assignment - pt2
A function that reads a directory full of files and reports the number 
of completely observed cases in each data file. The function should return a 
data frame where the first column is the name of the file and the second column
is the number of complete cases. A prototype of this function follows